agents/pyproject.toml
[project]
name = "metrosense-agents"
version = "0.1.0"
description = "MetroSense internal ADK agents service"
requires-python = ">=3.12"
dependencies = [
  "google-adk[eval]",
  "fastapi>=0.110",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "pydantic>=2.7",
  "uvicorn>=0.29",
]

[tool.uv]
managed = true
